+++ src/main/java/com/rearmhq/javaclient/interceptors/RearmBasicAuthInterceptor.java
@@ -1,4 +1,4 @@
-package rearm.java.client.interceptors;
+package com.rearmhq.javaclient.interceptors;
 
 import java.io.IOException;
